+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 - LDAPAuthenticator: New options to delete users and or groups on sync
+- Files: fixed migration error when folders had to be moved to new location
 
 29-03-2022 6.6.61
 - Projects: Refactored task editing from projects tasks tab